+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use top level `README.md` for `ogcapi` crate.
 - Rework features to represent modules and standards in an orthogonal fashion.
 - Convert CLI to example crates.
+- Align `OGCAPI - Processes` with released core standard.
 
 ## [0.2.0] - 2024-05-19